# tia/dua-tia-configuration-manager/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/tia/api/v1/query-max-publication-date": {
            "get": {
                "tags": [
                    "query-max-publication-date-controller"
                ],
                "operationId": "getQueryMaxPublicationDateList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QueryMaxPublicationDate"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "query-max-publication-date-controller"
                ],
                "operationId": "saveQueryMaxPublicationDate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QueryMaxPublicationDate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/QueryMaxPublicationDate"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/keyword-dictionaries": {
            "get": {
                "tags": [
                    "keyword-dictionary-controller"
                ],
                "operationId": "getKeywordDictionaryList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/KeywordDictionary"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "keyword-dictionary-controller"
                ],
                "operationId": "saveKeywordDictionary",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/KeywordDictionary"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/KeywordDictionary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/initiatives": {
            "get": {
                "tags": [
                    "initiative-controller"
                ],
                "operationId": "getInitiatives",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Initiative"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "initiative-controller"
                ],
                "operationId": "saveInitiative",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Initiative"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Initiative"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/tia/api/v1/query-max-publication-date/{searchBindingId}/{repositoryId}/{mission}": {
            "get": {
                "tags": [
                    "query-max-publication-date-controller"
                ],
                "operationId": "getQueryMaxPublicationDate",
                "parameters": [
                    {
                        "name": "searchBindingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "repositoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/QueryMaxPublicationDate"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "query-max-publication-date-controller"
                ],
                "operationId": "deleteQueryMaxPublicationDate",
                "parameters": [
                    {
                        "name": "searchBindingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "repositoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/query-max-publication-date/{id}": {
            "get": {
                "tags": [
                    "query-max-publication-date-controller"
                ],
                "operationId": "getQueryMaxPublicationDate_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/QueryMaxPublicationDate"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/keyword-dictionaries/{id}": {
            "get": {
                "tags": [
                    "keyword-dictionary-controller"
                ],
                "operationId": "getKeywordDictionary",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/KeywordDictionary"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "keyword-dictionary-controller"
                ],
                "operationId": "deleteKeywordDictionary",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tia/api/v1/initiatives/{id}": {
            "get": {
                "tags": [
                    "initiative-controller"
                ],
                "operationId": "getInitiative",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Initiative"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "initiative-controller"
                ],
                "operationId": "deleteInitiative",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "QueryMaxPublicationDate": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "searchBindingId": {
                        "type": "string"
                    },
                    "repositoryId": {
                        "type": "string"
                    },
                    "mission": {
                        "type": "string"
                    },
                    "maxPublicationDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Comments": {
                "type": "object",
                "properties": {
                    "generatedBy": {
                        "type": "string"
                    },
                    "generationTime": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "KeywordDictionary": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "shortName": {
                        "type": "string"
                    },
                    "get_comments": {
                        "$ref": "#/components/schemas/Comments"
                    },
                    "dictionaryDefinition": {
                        "$ref": "#/components/schemas/KeywordDictionaryDefinition"
                    }
                }
            },
            "KeywordDictionaryDefinition": {
                "type": "object",
                "properties": {
                    "keywords": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "skipKeywords": {
                        "type": "boolean"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/KeywordDictionaryDefinition"
                        }
                    }
                }
            },
            "Initiative": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "shortName": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}